In wfa_sock.c, line 415, the enum constant WFA_ERROR is returned instead of FALSE to fix the issue of SIG-465.
In wfa_tg.c, line 307-312, the assigned value to totalpkts variable is casted into int type to fix the issue related to SIG-523.

In wfa_tg.c, line 1862, the enum constant WFA_ERROR is returned instead of FALSE to fix the issue in SIG-465.

In wfa_thr.c, line 842, the logic handling the return value 0 is added because if the connection is gracefully closed, the return value is 0 and it needs to be handled, otherwise the infinite loop will occur as described in SIG-465.

SCR#SIG-465 and SIG-523
close[s:jira, t:465] close[s:jira, t:523]